chore: delete component refactoring
