+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@devtron-labs/devtron-fe-common-lib",
-    "version": "1.5.7-beta-7",
+    "version": "1.5.7-beta-8",
     "description": "Supporting common component library",
     "type": "module",
     "main": "dist/index.js",
