--- src/Shared/Components/ConfirmationModal/constants.tsx
@@ -1,4 +1,0 @@
-export enum ConfirmationActionType {
-    DELETE = 'delete',
-    CANCEL = 'cancel',
-}
